Continued refactor.
Cleanup of gene constructor class.
Precompiled regex before use. 
Using splitters from dcc-commons. 
Default description is now empty string rather than null.
