+++ dcc-import-gene/src/main/java/org/icgc/dcc/imports/gene/reader/ASNReader.java
@@ -87,7 +87,6 @@ public class ASNReader {
     val fOut = new FileOutputStream(tmpFile);
 
     ByteStreams.copy(gzip, fOut);
-
     return tmpFile.getAbsolutePath();
   }
